Rename feature to `easy-toggle-files` and update readme description
