+++ source/refined-github.ts
@@ -202,7 +202,7 @@ import './features/single-diff-column-selection';
 import './features/jump-to-change-requested-comment';
 import './features/esc-to-cancel';
 import './features/quick-fork-deletion';
-import './features/pr-easy-toggle-files';
+import './features/easy-toggle-files';
 
 // Add global for easier debugging
 (window as any).select = select;
